+++ src/views/dy/panorama/components/HalfPieChart.vue
